refactor(asusctl): modularize CLI handlers into dedicated modules and update mocking
