+++ asusctl/src/anime_cli.rs
@@ -1,5 +1,6 @@
 use argh::FromArgs;
-use rog_anime::usb::{AnimAwake, AnimBooting, AnimShutdown, AnimSleeping};
+use log::warn;
+use rog_anime::usb::{get_anime_type, AnimAwake, AnimBooting, AnimShutdown, AnimSleeping};
 use rog_anime::AnimeType;
 
 #[derive(FromArgs, Debug)]
@@ -148,4 +149,228 @@ pub struct AnimeGifDiagonal {
         description = "how many loops to play - 0 is infinite"
     )]
     pub loops: u32,
+}
+
+pub fn handle_anime(cmd: &AnimeCommand) -> Result<(), Box<dyn std::error::Error>> {
+    if cmd.command.is_none()
+        && cmd.enable_display.is_none()
+        && cmd.enable_powersave_anim.is_none()
+        && cmd.brightness.is_none()
+        && cmd.off_when_lid_closed.is_none()
+        && cmd.off_when_suspended.is_none()
+        && cmd.off_when_unplugged.is_none()
+        && cmd.off_with_his_head.is_none()
+        && !cmd.clear
+    {
+        warn!("Missing arg or command; run 'asusctl anime --help' for usage");
+        return Ok(());
+    }
+
+    let animes = crate::platform_cli::find_iface_blocking::<
+        rog_dbus::zbus_anime::AnimeProxyBlocking,
+    >("xyz.ljones.Anime")?;
+
+    for proxy in animes {
+        if let Some(enable) = cmd.enable_display {
+            proxy.set_enable_display(enable)?;
+        }
+        if let Some(enable) = cmd.enable_powersave_anim {
+            proxy.set_builtins_enabled(enable)?;
+        }
+        if let Some(bright) = cmd.brightness {
+            proxy.set_brightness(bright)?;
+        }
+        if let Some(enable) = cmd.off_when_lid_closed {
+            proxy.set_off_when_lid_closed(enable)?;
+        }
+        if let Some(enable) = cmd.off_when_suspended {
+            proxy.set_off_when_suspended(enable)?;
+        }
+        if let Some(enable) = cmd.off_when_unplugged {
+            proxy.set_off_when_unplugged(enable)?;
+        }
+
+        let mut anime_type = get_anime_type();
+        if let AnimeType::Unsupported = anime_type {
+            if let Some(model) = cmd.override_type {
+                anime_type = model;
+            }
+        }
+
+        if cmd.clear {
+            let data = vec![255u8; anime_type.data_length()];
+            let tmp = rog_anime::AnimeDataBuffer::from_vec(anime_type, data)?;
+            proxy.write(tmp)?;
+        }
+
+        if let Some(action) = cmd.command.as_ref() {
+            match action {
+                AnimeActions::Image(image) => {
+                    if image.path.is_empty() {
+                        warn!("Missing arg or command; run 'asusctl anime image --help' for usage");
+                        return Ok(());
+                    }
+                    verify_brightness(image.bright)?;
+
+                    let matrix = rog_anime::AnimeImage::from_png(
+                        std::path::Path::new(&image.path),
+                        image.scale,
+                        image.angle,
+                        rog_anime::Vec2::new(image.x_pos, image.y_pos),
+                        image.bright,
+                        anime_type,
+                    )?;
+
+                    proxy.write(<rog_anime::AnimeDataBuffer>::try_from(&matrix)?)?;
+                }
+                AnimeActions::PixelImage(image) => {
+                    if image.path.is_empty() {
+                        warn!("Missing arg or command; run 'asusctl anime pixel-image --help' for usage");
+                        return Ok(());
+                    }
+                    verify_brightness(image.bright)?;
+
+                    let matrix = rog_anime::AnimeDiagonal::from_png(
+                        std::path::Path::new(&image.path),
+                        None,
+                        image.bright,
+                        anime_type,
+                    )?;
+
+                    proxy.write(matrix.into_data_buffer(anime_type)?)?;
+                }
+                AnimeActions::Gif(gif) => {
+                    if gif.path.is_empty() {
+                        warn!("Missing arg or command; run 'asusctl anime gif --help' for usage");
+                        return Ok(());
+                    }
+                    verify_brightness(gif.bright)?;
+
+                    let matrix = rog_anime::AnimeGif::from_gif(
+                        std::path::Path::new(&gif.path),
+                        gif.scale,
+                        gif.angle,
+                        rog_anime::Vec2::new(gif.x_pos, gif.y_pos),
+                        rog_anime::AnimTime::Count(1),
+                        gif.bright,
+                        anime_type,
+                    )?;
+
+                    play_gif_animation(&proxy, &matrix, gif.loops)?;
+                }
+                AnimeActions::PixelGif(gif) => {
+                    if gif.path.is_empty() {
+                        warn!("Missing arg or command; run 'asusctl anime pixel-gif --help' for usage");
+                        return Ok(());
+                    }
+                    verify_brightness(gif.bright)?;
+
+                    let matrix = rog_anime::AnimeGif::from_diagonal_gif(
+                        std::path::Path::new(&gif.path),
+                        rog_anime::AnimTime::Count(1),
+                        gif.bright,
+                        anime_type,
+                    )?;
+
+                    play_gif_animation(&proxy, &matrix, gif.loops)?;
+                }
+                AnimeActions::SetBuiltins(builtins) => {
+                    if builtins.set.is_none() {
+                        warn!("Missing arg; run 'asusctl anime set-builtins --help' for usage");
+                        return Ok(());
+                    }
+
+                    proxy.set_builtin_animations(rog_anime::Animations {
+                        boot: builtins.boot,
+                        awake: builtins.awake,
+                        sleep: builtins.sleep,
+                        shutdown: builtins.shutdown,
+                    })?;
+                }
+            }
+        }
+    }
+    Ok(())
+}
+
+/// Helper to determine the playback iteration strategy.
+/// Returns `None` for infinite loops (`loops == 0`), or `Some(count)` for finite playback.
+fn compute_loop_plan(loops: u32) -> Option<u32> {
+    if loops == 0 {
+        None
+    } else {
+        Some(loops)
+    }
+}
+
+/// Play GIF animation frames. `loops == 0` means infinite playback until interrupted.
+fn play_gif_animation(
+    proxy: &rog_dbus::zbus_anime::AnimeProxyBlocking,
+    matrix: &rog_anime::AnimeGif,
+    loops: u32,
+) -> Result<(), Box<dyn std::error::Error>> {
+    let mut remaining = compute_loop_plan(loops);
+
+    loop {
+        for frame in matrix.frames() {
+            proxy.write(frame.frame().clone())?;
+            std::thread::sleep(frame.delay());
+        }
+        match remaining {
+            None => continue,
+            Some(1) => break,
+            Some(ref mut count) => *count -= 1,
+        }
+    }
+    Ok(())
+}
+
+fn verify_brightness(brightness: f32) -> Result<(), Box<dyn std::error::Error>> {
+    if !(0.0..=1.0).contains(&brightness) {
+        return Err(format!(
+            "Brightness must be between 0.0 and 1.0 (inclusive), was {brightness}"
+        )
+        .into());
+    }
+    Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_verify_brightness_valid() {
+        assert!(verify_brightness(0.0).is_ok());
+        assert!(verify_brightness(0.5).is_ok());
+        assert!(verify_brightness(1.0).is_ok());
+    }
+
+    #[test]
+    fn test_verify_brightness_invalid() {
+        assert!(verify_brightness(-0.1).is_err());
+        assert!(verify_brightness(1.1).is_err());
+        assert!(verify_brightness(f32::NAN).is_err());
+        assert!(verify_brightness(f32::INFINITY).is_err());
+        assert!(verify_brightness(f32::NEG_INFINITY).is_err());
+    }
+
+    #[test]
+    fn test_loop_iteration_count() {
+        assert_eq!(
+            compute_loop_plan(0),
+            None,
+            "0 loops must plan for infinite playback"
+        );
+        assert_eq!(
+            compute_loop_plan(1),
+            Some(1),
+            "1 loop must plan for 1 iteration"
+        );
+        assert_eq!(
+            compute_loop_plan(5),
+            Some(5),
+            "5 loops must plan for 5 iterations"
+        );
+    }
 }
